+++ src/main/java/com/altf4/app/model/monthlypayments/LoanTerms.java
@@ -1,4 +1,4 @@
-package com.altf4.app.model.calculation;
+package com.altf4.app.model.monthlypayments;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
@@ -9,7 +9,5 @@ public class LoanTerms {
 
     public static final double EURIBOR_RATE = 0.03335;
     public static final double MARGIN = 0.025;
-    public static final double LOAN_SERVICE_RATIO = 0.4;
-    public static final double LOAN_TO_DOWN_PAYMENT_RATIO = 0.85;
-}
 
+}
