+++ src/main/java/com/altf4/app/service/LoanCalculatorService.java
@@ -1,23 +1,23 @@
 package com.altf4.app.service;
 
-import com.altf4.app.model.calculation.LoanCalculationRequest;
-import com.altf4.app.model.calculation.LoanCalculationResponse;
+import com.altf4.app.model.monthlypayments.MonthlyPaymentsRequest;
+import com.altf4.app.model.monthlypayments.MonthlyPaymentsResponse;
 import org.springframework.stereotype.Service;
 
-import static com.altf4.app.model.calculation.LoanTerms.EURIBOR_RATE;
-import static com.altf4.app.model.calculation.LoanTerms.MARGIN;
+import static com.altf4.app.model.monthlypayments.LoanTerms.EURIBOR_RATE;
+import static com.altf4.app.model.monthlypayments.LoanTerms.MARGIN;
 
 @Service
 public class LoanCalculatorService {
 
 
-    public LoanCalculationResponse getLoanCalculations(LoanCalculationRequest request) {
+    public MonthlyPaymentsResponse getLoanCalculations(MonthlyPaymentsRequest request) {
 
         int monthlyPayment = calculateMonthlyPayment(request);
         return buildResponse(request, monthlyPayment);
     }
 
-    private static int calculateMonthlyPayment(LoanCalculationRequest request) {
+    private static int calculateMonthlyPayment(MonthlyPaymentsRequest request) {
         int banksLoanAmount = calculateLoanAmount(request);
         double monthlyInterestRate = (EURIBOR_RATE + MARGIN) / 12;
         int loanTermMonths = request.getTermYears() * 12;
@@ -27,8 +27,8 @@ public class LoanCalculatorService {
                 / (Math.pow(1 + monthlyInterestRate, loanTermMonths) - 1) / 10.0) * 10);
     }
 
-    private static LoanCalculationResponse buildResponse(LoanCalculationRequest request, int monthlyPayment) {
-        return new LoanCalculationResponse.LoanCalculationResponseBuilder()
+    private static MonthlyPaymentsResponse buildResponse(MonthlyPaymentsRequest request, int monthlyPayment) {
+        return new MonthlyPaymentsResponse.LoanCalculationResponseBuilder()
                 .monthlyPaymentAmount(monthlyPayment)
                 .interestRate(EURIBOR_RATE + MARGIN)
                 .loanAmount(calculateLoanAmount(request))
@@ -37,15 +37,15 @@ public class LoanCalculatorService {
                 .build();
     }
 
-    private static int calculateLoanAmount(LoanCalculationRequest request) {
+    private static int calculateLoanAmount(MonthlyPaymentsRequest request) {
         return request.getTotalAmount() - request.getDownPayment();
     }
 
-    private static int calculateTotalInterestAmount(LoanCalculationRequest request, int monthlyPayment) {
+    private static int calculateTotalInterestAmount(MonthlyPaymentsRequest request, int monthlyPayment) {
         return monthlyPayment * 12 * request.getTermYears() - (request.getTotalAmount() - request.getDownPayment());
     }
 
-    private static int calculateTotalPaymentSum(LoanCalculationRequest request, int monthlyPayment) {
+    private static int calculateTotalPaymentSum(MonthlyPaymentsRequest request, int monthlyPayment) {
         return calculateLoanAmount(request) + calculateTotalInterestAmount(request, monthlyPayment);
     }
