+++ src/main/java/com/altf4/app/util/StringToEnumConverter.java
