+++ src/main/java/com/altf4/app/validator/FinancialInformationConstraint.java
