+++ src/main/java/com/altf4/app/validator/ValidationsConstants.java
@@ -11,4 +11,5 @@ public class ValidationsConstants {
     public static final int MAXIMUM_LOAN_TERM = 30;
 
     public static final double DOWN_PAYMENT_TO_LOAN_RATIO = 0.15;
+
 }
